create calculator component and organize folders
