+++ src/components/Calculator/Calculator.css
@@ -1,0 +1,3 @@
+.Calculator {
+  padding: 20px;
+}
