+++ src/routes/home/index.tsx
@@ -1,23 +1,11 @@
 import { h } from 'preact';
-import Interpreter from '../../components/Interpreter';
+import Calculator from '../../components/Calculator/Calculator';
 
 // TODO support css modules
 const style = require('./style');
 
 const Home = (_props: any) => {
-  return (
-    <div class={style.home}>
-      <Interpreter value={'2+2'} />
-      <Interpreter value={'5-10'} />
-      <Interpreter value={'100*4'} />
-      <Interpreter value={'100/4'} />
-      <Interpreter value={'1*2+3'} />
-      <Interpreter value={'1+2*3'} />
-      <Interpreter value={'(1+2)*3'} />
-      <Interpreter value={'1*(1+2)*3+(1+2+3)'} />
-      <Interpreter value={')1+2)*3'} />
-    </div>
-  );
+  return <Calculator />;
 };
 
 export default Home;
